requirements.txt
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
httpx==0.27.2
openai==1.30.1
pytest==8.3.5
requests==2.32.3
